The program is at the point where bugs need to be fixed.
Adjusted ROI so the top left of the box lands exactly where the mouse was clicked.

To Shane:
I can't figure out why the image starts off being offset (it's down and to the right of where the image actually starts), if you could fix that (or figure why it is happening), that would be much appreciated.
-Michael